# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Freelance & Client Job API",
    "version": "1.0.0",
    "description": "API for authentication, user management, and client job CRUD with JWT + RBAC"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Job": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "budget": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              },
              "currency": {
                "type": "string"
              }
            }
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "experienceLevel": {
            "type": "string"
          },
          "isUrgent": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/jobs/create": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create new job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "skills",
                  "minBudget",
                  "maxBudget",
                  "deadline"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "skills": {
                    "type": "string",
                    "description": "Comma-separated"
                  },
                  "minBudget": {
                    "type": "number"
                  },
                  "maxBudget": {
                    "type": "number"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "experienceLevel": {
                    "type": "string"
                  },
                  "isUrgent": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}